requirements.txt
packaging
Pillow